% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptNextPoint.R
\name{acceptNextPoint}
\alias{acceptNextPoint}
\title{Accept next point ?}
\usage{
acceptNextPoint(current.density, proposed.density)
}
\arguments{
\item{current.density}{Numeric scalar, density at the current point.}

\item{proposed.density}{Numeric scalar, density at the proposed point.}
}
\value{
TRUE if the proposed point should be accepted
}
\description{
\code{acceptNextPoint} checks whether the proposed point should be accepted
}
\keyword{internal}
