% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHighDimGaussian.R
\name{addHighDimGaussian}
\alias{addHighDimGaussian}
\title{addHighDimGaussian}
\usage{
addHighDimGaussian(
  dim = 0,
  mean.vec = matrix(0, ncol = dim),
  cov.mat = diag(dim)
)
}
\arguments{
\item{dim}{integer that specifies the number of dimensions that should be altered}

\item{mean.vec}{vector of the means of the Gaussian to be added}

\item{cov.mat}{covariance matrix of the Gaussian to be added}
}
\value{
function that takes a point given as a dataframe as input and returns it with Gaussian noise added to the specified dimensions
}
\description{
Adds high dimensional Gaussian noise to specific parameters of a point given as a dataframe.
}
