% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alwaysOne.R
\name{alwaysOne}
\alias{alwaysOne}
\title{AlwaysOne}
\usage{
alwaysOne(...)
}
\arguments{
\item{...}{Any input is possible, as it does not have any effect}
}
\value{
1, always
}
\description{
As the name implies, this function returns always 1, no matter the input. It can be used as a dummy desity function to perform uniform sampling of higher dimensional space. As it works with any Input, it can also be used as a placeholder to evade errors.
}
\keyword{alwaysOne(stop("even}
\keyword{fine"))}
\keyword{input}
\keyword{internal}
\keyword{is}
\keyword{this}
