% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-helpers.R
\name{check_columns_exist}
\alias{check_columns_exist}
\title{Check that column names exist in a dataset}
\usage{
check_columns_exist(
  data,
  columns,
  data_name = "dataset",
  col_name = "dimensions"
)
}
\arguments{
\item{data}{The data.frame or sf object to check columns in}

\item{columns}{Character vector of column names to look for}

\item{data_name}{Name of the data argument (for error messages)}

\item{col_name}{Name of the columns argument (for error messages)}
}
\description{
Check that column names exist in a dataset
}
\keyword{internal}
