% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-helpers.R
\name{check_in_range}
\alias{check_in_range}
\title{Check that numeric values fall within a range}
\usage{
check_in_range(x, arg_name, min_val = 0, max_val = 1)
}
\arguments{
\item{x}{The value(s) to check}

\item{arg_name}{Name of the argument (for error messages)}

\item{min_val}{Minimum allowed value (inclusive)}

\item{max_val}{Maximum allowed value (inclusive)}
}
\description{
Check that numeric values fall within a range
}
\keyword{internal}
