% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-helpers.R
\name{check_positive_integer}
\alias{check_positive_integer}
\title{Check that a value is a positive integer}
\usage{
check_positive_integer(x, arg_name, allow_zero = FALSE)
}
\arguments{
\item{x}{The value to check}

\item{arg_name}{Name of the argument (for error messages)}

\item{allow_zero}{If TRUE, zero is accepted}
}
\description{
Check that a value is a positive integer
}
\keyword{internal}
