% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-helpers.R
\name{check_raster_input}
\alias{check_raster_input}
\title{Check that a raster input is valid}
\usage{
check_raster_input(x, arg_name = "env.rast")
}
\arguments{
\item{x}{The object to check}

\item{arg_name}{Name of the argument (for error messages)}
}
\description{
Check that a raster input is valid
}
\keyword{internal}
