% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-helpers.R
\name{check_spatial_points}
\alias{check_spatial_points}
\title{Check that a spatial points input is valid}
\usage{
check_spatial_points(x, arg_name = "pres", allow_null = FALSE)
}
\arguments{
\item{x}{The object to check}

\item{arg_name}{Name of the argument (for error messages)}

\item{allow_null}{If TRUE, NULL values are accepted without error}
}
\description{
Check that a spatial points input is valid
}
\keyword{internal}
