% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapBackOnRealPoints.R
\name{mapBackOnRealPoints}
\alias{mapBackOnRealPoints}
\title{Map Back on Real Points
searches the closest point in the dataset regarding the given point and the dimensions given}
\usage{
mapBackOnRealPoints(dataset = NULL, point = NULL, dim = "", threshold = 0.5)
}
\arguments{
\item{dataset}{dataframe of the target dataset}

\item{point}{dataframe containing the given point}

\item{dim}{string vector containing the names of the dimensions that should be respected}

\item{threshold}{threshold obove which we consider the distance too big and want to discard the point}
}
\value{
closest point that we could find in the dataset with the distance added as a parameter
}
\description{
Map Back on Real Points
searches the closest point in the dataset regarding the given point and the dimensions given
}
