% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWithLines.R
\name{plot_points_with_lines}
\alias{plot_points_with_lines}
\title{Plott a dequence of points in 2d}
\usage{
plot_points_with_lines(
  df,
  cols,
  limits = NULL,
  title = "Connected Data Points"
)
}
\arguments{
\item{df}{dataframe containing the observation}

\item{cols}{vector containing the names of the columns to be plotted}

\item{limits}{sets the limits of the plot. This is usefull to ensure comparable plot sizes}

\item{title}{title of the plot}
}
\value{
the desired plot
}
\description{
Plott a dequence of points in 2d
}
