% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensity.R
\name{plotDensity2dpro}
\alias{plotDensity2dpro}
\title{Function to plot the density of a density}
\usage{
plotDensity2dpro(
  dataset,
  species = NULL,
  xlim = c(0, 1),
  ylim = c(0, 1),
  densityFunction = NULL,
  resolution = 10
)
}
\arguments{
\item{dataset}{dataframe containing points that can be added to the desity field. The parameter is needed as the code
uses it as a template for points. If its size is greater than one the points are added to the plot}

\item{species}{dataframe containing the species points to be added to the plot}

\item{xlim}{Limits of the plot}

\item{ylim}{Limits of the plot}

\item{densityFunction}{Density to be plottet. The function should take a point as its argument and give back a float}

\item{resolution}{number of gridcells in each direction}
}
\value{
list containing the sampling grid as well as the density matrix
}
\description{
Function to plot the density of a density
}
