% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensityLines.R
\name{plotDensityLines}
\alias{plotDensityLines}
\title{plotDensityLines
enables the plotting of density function as well as a trace of a chain made of points in a dataframe
on this density surface. In addition the species that was used to generate the density model can be supplied
to verify the performance of the model. If that is the case the supplied dataset will be plotted as points
instead of as a chain. This function has some issues with the check of devtools}
\usage{
plotDensityLines(
  dataset,
  xlim = c(0, 1),
  ylim = c(0, 1),
  title = "Connected Data Points",
  cols = NULL,
  lines = FALSE,
  density = FALSE,
  species = NULL,
  densityFunction = alwaysOne,
  resolution = 10,
  minimal = FALSE
)
}
\arguments{
\item{dataset}{Dataframe specifying the dataset to be plotted as a line or points respectively}

\item{xlim}{x-limit of the plot}

\item{ylim}{y-limit of the plot}

\item{title}{title of the plot}

\item{cols}{specifies the columns of the dataframe that are used for the density computation. It has to match to the columns used to build the density model.}

\item{lines}{boolean that tells the function if the lines should be plotted. If false the dataset will be ploted as points}

\item{density}{boolean that tells the function if the density should be plotted}

\item{species}{dataframe containing the simulated species, it has to contain the column names of cols}

\item{densityFunction}{a function that can take a dataframe with the columns given in cols and retruns the density at that point}

\item{resolution}{int sets the resolution of the denisity grid}

\item{minimal}{boolean if true removes titel, label, legend and axis}
}
\value{
the greated plot
}
\description{
plotDensityLines
enables the plotting of density function as well as a trace of a chain made of points in a dataframe
on this density surface. In addition the species that was used to generate the density model can be supplied
to verify the performance of the model. If that is the case the supplied dataset will be plotted as points
instead of as a chain. This function has some issues with the check of devtools
}
