% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptNextPoint.R
\name{acceptNextPoint}
\alias{acceptNextPoint}
\title{Accept next point ?}
\usage{
acceptNextPoint(current.point, proposed.point, densityFunction)
}
\arguments{
\item{current.point}{Current state of the markov chain, given as a sf dataframe.}

\item{proposed.point}{Proposed next state of the markov chain, given as a sf dataframe.}

\item{densityFunction}{Function that can take a point given as a sf dataframe as a input and returns the target density at that location.}
}
\value{
TRUE if the proposed point should be accepted
}
\description{
\code{acceptNextPoint} checks whether the proposed point should be accepted
}
\keyword{internal}
