% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHighDimGaussian.R
\name{addHighDimGaussian}
\alias{addHighDimGaussian}
\title{addHighDimGaussian}
\usage{
addHighDimGaussian(
  currentPoint,
  dim = "",
  mean_vec = matrix(0, ncol = length(dim)),
  cov_mat = 2.3 * diag(length(dim))
)
}
\arguments{
\item{currentPoint}{Current point as a dataframe from which a new point should be generated}

\item{dim}{vector with the names of the columns on which we want to add}

\item{mean_vec}{vector of the means of the Gaussian to be added}

\item{cov_mat}{covariance matrix of the Gaussian to be added}
}
\value{
dataframe with Gaussian noise added to the specified columns
}
\description{
Adds high dimensional Gaussian noise to specific parameters of a point given as a dataframe.
}
\keyword{internal}
