% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHighDimGaussian.R
\name{addHighDimGaussian}
\alias{addHighDimGaussian}
\title{addHighDimGaussian}
\usage{
addHighDimGaussian(
  dim = 0,
  mean_vec = matrix(0, ncol = dim),
  cov_mat = diag(dim)
)
}
\arguments{
\item{dim}{integer that specifies the number of dimensions that should be altered}

\item{mean_vec}{vector of the means of the Gaussian to be added}

\item{cov_mat}{covariance matrix of the Gaussian to be added}
}
\value{
a function that takes a point given as a dataframe as input and returns it with gaussian noise added to the specified dimensions
}
\description{
Adds high dimensional Gaussian noise to specific parameters of a point given as a dataframe.
}
