% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuclidianMetric.R
\name{euclidianMetric}
\alias{euclidianMetric}
\title{Euclidean metric on specified columns}
\usage{
euclidianMetric(point.a = NULL, point.b = NULL, dim = "")
}
\arguments{
\item{point.a}{First point}

\item{point.b}{Second point}

\item{dim}{vector containing the names of the columns that should be included in the computation}
}
\value{
euclidean metric of the two points
}
\description{
\code{euclidianMetric} calculates the euclidean metric of specified columns of two points given as dataframes
}
\keyword{internal}
