% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVirtualSpeciesPrecencePoints.R
\name{getVirtualSpeciesPresencePoints}
\alias{getVirtualSpeciesPresencePoints}
\title{Function to streamline the generation and sampling of a virtual species}
\usage{
getVirtualSpeciesPresencePoints(env.data = NULL, n.samples = 0)
}
\arguments{
\item{env.data}{Raster containing environmental parameters}

\item{n.samples}{Number of samples}
}
\value{
List containing the sampled points as a dataframe as well as other nice things that can later be used to plot
}
\description{
Function to streamline the generation and sampling of a virtual species
}
