% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcSampling.R
\name{mcmcSampling}
\alias{mcmcSampling}
\title{MCMC sampling from a given dataset}
\usage{
mcmcSampling(
  dataset = NULL,
  dimensions = list(""),
  densityFunction = alwaysOne,
  proposalFunction = addHighDimGaussian(dim = length(dimensions)),
  n.sample.points = 0,
  burnIn = 1000,
  verbose = TRUE,
  covariance.correction = 1,
  max.burnin.cycles = 50
)
}
\arguments{
\item{dataset}{sf dataframe from which the points are sampled}

\item{dimensions}{string vector containing the dimensions that should be included in the random walk}

\item{densityFunction}{Function that can take a point given as a numeric vector as input and returns the target density at that location.}

\item{proposalFunction}{Function that can take a point given as a numeric vector and a covariance adjuster as input and returns a new proposed point as a numeric vector.}

\item{n.sample.points}{Number of points to be sampled}

\item{burnIn}{Integer, sets the number of samples per adaptive burn in step. If set to 0, burn in is skipped}

\item{verbose}{Boolean to toggle progress updates}

\item{covariance.correction}{Integer, initial value of the covariance correction.}

\item{max.burnin.cycles}{Integer, maximum number of burn-in adaptation cycles before stopping with a warning. Prevents infinite loops when the target acceptance rate cannot be reached.}
}
\value{
A data.frame containing the sampled points with dimension columns and a density column
}
\description{
MCMC sampling from a given dataset
}
