% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcSampling.R
\name{mcmcSampling}
\alias{mcmcSampling}
\title{MCMC sampling from a given dataset}
\usage{
mcmcSampling(
  dataset = NULL,
  dimensions = list(""),
  densityFunction = alwaysOne,
  proposalFunction = addHighDimGaussian(dim = length(dimensions)),
  n.sample.points = 0
)
}
\arguments{
\item{dataset}{sf dataframe from which the points are sampled}

\item{dimensions}{string vector containing the dimensions that should be included in the random walk}

\item{densityFunction}{Function that can take a point given as a sf dataframe as a input and returns the target density at that location.}

\item{proposalFunction}{Function that can take a point given as a sf dataframe and a vector of strings specifying the row names that should be changed as a input and returns a new proposed point}

\item{n.sample.points}{Number of points to be sampled}
}
\value{
A sf dataframe containing the sampled points
}
\description{
MCMC sampling from a given dataset
}
