% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInGeographicalSpace.R
\name{plotInGeographicalSpace}
\alias{plotInGeographicalSpace}
\title{Function that plots the geographical location of points onto a raster}
\usage{
plotInGeographicalSpace(
  presence.distribution.raster = NULL,
  presence.points = NULL,
  absence.points = NULL,
  minimal = FALSE
)
}
\arguments{
\item{presence.distribution.raster}{Raster containing the target distribution}

\item{presence.points}{Dataframe or list containing a geometry cloumn that contains points to be plotted}

\item{absence.points}{Dataframe or list containing a geometry cloumn that contains points to be plotted}

\item{minimal}{Boolean if TRUE removes the labels legend and title}
}
\value{
NULL, as the function just plots
}
\description{
Function that plots the geographical location of points onto a raster
}
