% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USE.R
\docType{package}
\name{USE.MCMC-package}
\alias{USE.MCMC}
\alias{USE.MCMC-package}
\title{USE (Uniform Sampling of the Environmental space)
An R package to perform uniform sampling in the environmental space}
\description{
USE.MCMC performs MCMC sampling of the environmental space.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mdmdma.github.io/USE.MCMC/}
  \item Report bugs at \url{https://github.com/mdmdma/USE.MCMC/issues}
}

}
\author{
\strong{Maintainer}: Mathis Erler \email{m.erler@gmx.ch}

Authors:
\itemize{
  \item Mathis Erler \email{m.erler@gmx.ch}
  \item Daniele Da Re \email{dare.daniele@gmail.com} (\href{https://orcid.org/0000-0002-3398-9295}{ORCID})
  \item Enrico Tordoni \email{e.tordoni@gmail.com} (\href{https://orcid.org/0000-0002-9722-6692}{ORCID})
  \item Manuele Bazzichetto \email{manuele.bazzichetto@gmail.com} (\href{https://orcid.org/0000-0002-9874-5064}{ORCID})
}

}
